requirements.txt
certifi==2018.1.18
chardet==3.0.4
future==0.16.0
idna==2.6
python-telegram-bot==9.0.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
